requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
boto3==1.20.23
botocore==1.23.24
Collectfast==2.2.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0
django-appconf==1.0.5
django-autocomplete-light==3.8.2
django-crispy-forms==1.13.0
django-environ==0.8.1
django-import-export==2.7.0
django-select2==7.9.0
django-selectable==1.3.0
django-smart-selects==1.5.9
django-storages==1.12.3
djangorestframework==3.13.0
et-xmlfile==1.1.0
gunicorn==20.1.0
jmespath==0.10.0
MarkupPy==1.14
numpy==1.21.4
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.3.4
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycodestyle==2.8.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
toml==0.10.2
typing-extensions==4.0.1
Unipath==1.1
urllib3==1.26.7
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
